+++ src/main/java/com/mathpar/parallel/dap/cholesky/MatrixS/Tests/MatrSCholFactStrassWin7Test.java
@@ -25,7 +25,7 @@ public class MatrSCholFactStrassWin7Test extends DAPTest {
     }
 
     @Override
-    protected MatrixS[] initData(int size, int density, int maxBits, Ring ring) {
+    protected MatrixS[] initData(int size, double density, int maxBits, Ring ring) {
         return new MatrixS[]{matrix(size, density, maxBits, ring)};
     }
 
@@ -55,7 +55,7 @@ public class MatrSCholFactStrassWin7Test extends DAPTest {
     }
 
     @Override
-    protected MatrixS matrix(int size, int density, int maxBits, Ring ring){
+    protected MatrixS matrix(int size, double density, int maxBits, Ring ring){
         MatrixS matrix = new MatrixS(size, size, density, new int[] {maxBits}, new Random(), ring.numberONE(), ring);
         for (int i = 0; i < size; i++) {
             for (int j = 0; j < size; j++) {
