+++ src/main/java/com/mathpar/parallel/dap/multiply/MatrixS/Tests/MatrSMult4Test.java
@@ -21,7 +21,7 @@ public class MatrSMult4Test extends DAPTest {
     }
 
     @Override
-    protected Element[] initData(int size, int density, int maxBits, Ring ring) {
+    protected Element[] initData(int size, double density, int maxBits, Ring ring) {
         MatrixS A = matrix(size, density, maxBits, ring);
         MatrixS B = matrix(size, density, maxBits, ring);
         return new Element[]{A, B};
@@ -52,7 +52,7 @@ public class MatrSMult4Test extends DAPTest {
     }
 
     @Override
-    protected MatrixS matrix(int size, int density, int maxBits, Ring ring){
+    protected MatrixS matrix(int size, double density, int maxBits, Ring ring){
         MatrixS matrix = new MatrixS(size, size, density, new int[]{maxBits}, new Random(),ring.numberONE(), ring);
         // LOGGER.trace("bef matrix = " + matrix);
         for (int i = 0; i < size; i++) {
