+++ src/main/java/com/mathpar/parallel/dap/test/MultiplyTest.java
@@ -26,7 +26,7 @@ public class MultiplyTest extends DAPTest {
     }
 
     @Override
-    protected MatrixS[] initData(int size, int density, int maxBits, Ring ring) {
+    protected MatrixS[] initData(int size, double density, int maxBits, Ring ring) {
         return new MatrixS[]{matrix(size, density, maxBits, ring), matrix(size, density, maxBits, ring)};
     }
 
@@ -55,7 +55,7 @@ public class MultiplyTest extends DAPTest {
     }
 
     @Override
-    protected MatrixS matrix(int size, int density, int maxBits, Ring ring){
+    protected MatrixS matrix(int size, double density, int maxBits, Ring ring){
         MatrixS matrix = new MatrixS(size, size, density, new int[]{maxBits}, new Random(),ring.numberONE(), ring);
        // LOGGER.trace("bef matrix = " + matrix);
         for (int i = 0; i < size; i++) {
