+++ src/main/java/com/mathpar/parallel/dap/test/QRDecompositionTest.java
@@ -27,7 +27,7 @@ public class QRDecompositionTest extends DAPTest {
     }
 
     @Override
-    protected MatrixS[] initData(int size, int density, int maxBits, Ring ring) {
+    protected MatrixS[] initData(int size, double density, int maxBits, Ring ring) {
         return new MatrixS[]{matrix(size, density, maxBits, ring)};
     }
